lwsadmin/requirements.txt
aiofiles==25.1.0
bcrypt==3.2.2
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
Hypercorn==0.18.0
hyperframe==6.1.0
idna==3.11
ipaddress==1.0.23
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
monero==1.1.1
peewee==3.15.3
pillow==12.1.0
priority==2.0.0
pycparser==2.23
pycryptodomex==3.23.0
PyNaCl==1.6.2
PySocks==1.7.1
python-dotenv==1.0.0
quart==0.18.3
quart-auth==0.8.0
quart-bcrypt==0.0.6
qrcode==8.2
requests==2.29.0
urllib3==1.26.20
varint==1.0.2
Werkzeug==2.2.2
wsproto==1.3.2
